requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
yt-dlp>=2023.12.30
pysrt>=1.1.2
pydub>=0.25.1
youtube-transcript-api>=0.6.1
tqdm>=4.66.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0